requirements.txt
asgiref==3.6.0
attrs==23.1.0
autoflake==2.2.0
autopep8==2.0.2
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
djando==0.2
Django==4.1.5
django-cors-headers==3.14.0
django-environ==0.9.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular
idna==3.4
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.8.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
